Finished Accounts pages

+++ PanGainsWebApp/Controllers/AccountsController.cs
@@ -107,10 +107,7 @@ namespace PanGainsWebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int accountID, [Bind("AccountID,Firstname,Lastname,Email,Password,Title,ProfilePicture,Description,Private,Notifications,AverageChallengePos,Type,Role")] Account account)
         {
-            if (accountID != account.AccountID)
-            {
-                return NotFound();
-            }
+            if (accountID != account.AccountID) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -138,7 +135,7 @@ namespace PanGainsWebApp.Controllers
         // GET: Accounts/Delete/5
         public async Task<IActionResult> Delete(int? accountID)
         {
-            if (accountID == null) return NotFound(); //error checking
+            if (accountID == null) return RedirectToAction(nameof(Index)); //error checking
 
             AccountDetails accountDetails = new AccountDetails();
 
@@ -153,7 +150,7 @@ namespace PanGainsWebApp.Controllers
             accountDetails.Followers = socialList.Where(s => s.FollowingID == accountID).ToList().Count();
             accountDetails.Following = socialList.Where(s => s.AccountID == accountID).ToList().Count();
 
-            if (accountDetails == null) return NotFound(); //error checking
+            if (accountDetails == null) return RedirectToAction(nameof(Index)); //error checking
 
             return View(accountDetails);
         }
@@ -163,9 +160,55 @@ namespace PanGainsWebApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int accountID)
         {
-            var account = await _context.Account.FirstOrDefaultAsync(a => a.AccountID == accountID);
-            _context.Account.Remove(account);
-            
+            //remove Statistics
+            if (_context.Statistics != null)
+            {
+                var statisticsList = await _context.Statistics.ToListAsync();
+                List<Statistics> statistics = statisticsList.Where(s => s.AccountID == accountID).ToList();
+                if (statistics.Any()) foreach (Statistics s in statistics) _context.Statistics.Remove(s);
+            }
+            //remove DaysWorkedOut
+            if (_context.DaysWorkedOut != null)
+            {
+                var daysWorkedOutsList = await _context.DaysWorkedOut.ToListAsync();
+                List<DaysWorkedOut> daysWorkedOuts = daysWorkedOutsList.Where(d => d.AccountID == accountID).ToList();
+                if (daysWorkedOuts.Any()) foreach (DaysWorkedOut d in daysWorkedOuts) _context.DaysWorkedOut.Remove(d);
+            }
+            //remove ChallengeStats
+            if (_context.ChallengeStats != null)
+            {
+                var challengeStatsList = await _context.ChallengeStats.ToListAsync();
+                List<ChallengeStats> challengeStats = challengeStatsList.Where(c => c.AccountID == accountID).ToList();
+                if (challengeStats.Any()) foreach (ChallengeStats c in challengeStats) _context.ChallengeStats.Remove(c);
+            }
+            //remove Social
+            if (_context.Social != null)
+            {
+                var socialsList = await _context.Social.ToListAsync();
+                List<Social> socials = socialsList.Where(s => s.AccountID == accountID).ToList();
+                if (socials.Any()) foreach (Social s in socials) _context.Social.Remove(s);
+            }
+            //remove CompletedWorkout
+            if (_context.CompletedWorkout != null)
+            {
+                var completedWorkoutsList = await _context.CompletedWorkout.ToListAsync();
+                List<CompletedWorkout> completedWorkouts = completedWorkoutsList.Where(c => c.AccountID == accountID).ToList();
+                if (completedWorkouts.Any()) foreach (CompletedWorkout c in completedWorkouts) _context.CompletedWorkout.Remove(c);
+            }
+            //remove Folder
+            if (_context.Folder != null)
+            {
+                var foldersList = await _context.Folder.ToListAsync();
+                List<Folder> folders = foldersList.Where(f => f.AccountID == accountID).ToList();
+                if (folders.Any()) foreach (Folder f in folders) _context.Folder.Remove(f);
+            }
+            //remove Account
+            if (_context.Account != null)
+            {
+                var account = await _context.Account.FindAsync(accountID);
+                if (account == null) return NotFound();
+                _context.Account.Remove(account);
+            }
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
